move chunk.ts file to core/tree
