+++ packages/dds/tree/src/core/tree/index.ts
@@ -110,5 +110,13 @@ export {
 	emptyDelta,
 } from "./deltaUtil.js";
 
+export {
+	type TreeChunk,
+	dummyRoot,
+	cursorChunk,
+	tryGetChunk,
+	type ChunkedCursor,
+} from "./chunk.js";
+
 export { DetachedFieldIndex } from "./detachedFieldIndex.js";
 export { type ForestRootId } from "./detachedFieldIndexTypes.js";
