+++ packages/dds/tree/src/feature-libraries/chunked-forest/chunkTree.ts
@@ -21,12 +21,13 @@ import {
 	mapCursorFields,
 	Multiplicity,
 	ValueSchema,
+	type TreeChunk,
+	tryGetChunk,
 } from "../../core/index.js";
 import { fail, getOrCreate } from "../../util/index.js";
 import type { FullSchemaPolicy } from "../modular-schema/index.js";
 
 import { BasicChunk } from "./basicChunk.js";
-import { type TreeChunk, tryGetChunk } from "./chunk.js";
 import { SequenceChunk } from "./sequenceChunk.js";
 import { type FieldShape, TreeShape, UniformChunk } from "./uniformChunk.js";
 import { isStableNodeKey } from "../node-key/index.js";
